requirements.txt
djangorestframework>=3.14
asgiref             3.8.1
Django              5.1.3
djangorestframework 3.15.2
pillow              11.0.0
pip                 24.2
sqlparse            0.5.2
tzdata              2024.2
stripe>=5.0
